requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.2.4
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
narwhals==2.16.0
numpy==2.0.2
oauthlib==3.3.1
openai==2.21.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.5
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pypdf==6.7.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
